+++ src/bittytax/conv/parsers/__init__.py
@@ -17,6 +17,7 @@ from . import (
     blockscout,
     bnktothefuture,
     bybit,
+    cashapp,
     celsius,
     cexio,
     cgtcalculator,