+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - Conversion tool: added parser for Robinhood exchange.
 - Conversion tool: added parser for Aptoscan explorer.
 - Conversion tool: added parser for FatStx explorer.
+- Conversion tool: added parser for Cash App wallet.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))